requirements.txt
annotated-types==0.7.0
anyio==4.9.0
appdirs==1.4.4
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
environs==14.1.1
fastapi==0.115.12
fastapi_cors==0.0.6
h11==0.14.0
httpcore==1.0.8
httpx==0.26.0
idna==3.10
Jinja2==3.1.6
lseg-data==2.1.1
MarkupSafe==3.0.2
marshmallow==4.0.0
numpy==1.26.4
pandas==2.2.3
pydantic==2.11.3
pydantic_core==2.33.1
pyee==11.1.0
pyhumps==3.8.0
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
refinitiv-data==1.6.2
requests==2.32.3
scipy==1.12.0
setuptools==79.0.1
simplejson==3.20.1
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
tenacity==8.5.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.2
watchdog==2.3.1
websocket-client==1.8.0
websockets==15.0.1
